sched/fair: Don't remove important task migration logic from PELT
This task migration logic is guarded by a WALT #ifdef even though it has
nothing specific to do with WALT. The result is that, with PELT, boosted
tasks can be migrated to the little cluster, causing visible stutters.

Move the WALT #ifdef so PELT can benefit from this logic as well.

Thanks to Zachariah Kennedy <zkennedy87@gmail.com> and Danny Lin
<danny@kdrag0n.dev> for discovering this issue and creating this fix.

Change-Id: Ia4c635463908e8738f21f8bce9c02ee8c8365bdf
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>